Improve draw buffer flow
